refactor: major cleanup + wiring up JSDOM replacement to puppeteer for SSR
