+++ packages/servers/ssr-server/package.json
@@ -5,13 +5,22 @@
   "license": "MIT",
   "module": "index.mjs",
   "scripts": {
+    "postbootstrap": "patch-package --patch-dir .patches",
+    "postinstall": "patch-package --patch-dir .patches",
     "start": "node --experimental-modules --quiet index.mjs"
   },
   "publishConfig": {
     "access": "public"
   },
   "dependencies": {
+    "cli-highlight": "^2.0.0", 
+    "document-register-element": "^1.13.1",
+    "express": "^4.16.3",
     "globby": "^9.0.0",
-    "express": "^4.16.3"
+    "is-reachable": "^3.0.0",
+    "mutationobserver-shim": "^0.3.3",
+    "patch-package": "^6.0.5",
+    "postinstall-postinstall": "^2.0.0",
+    "raf": "^3.4.1"
   }
 }
